Switched to using dictionary
