--- maps/v1/dict.go
@@ -1,7 +1,0 @@
-package main
-
-type Dict map[string]string
-
-func (d Dict) Search(word string) string {
-	return d[word]
-}
