+++ maps/v1/dictionary.go
@@ -1,0 +1,7 @@
+package main
+
+type Dictionary map[string]string
+
+func (d Dictionary) Search(word string) string {
+	return d[word]
+}
