+++ maps/v4/dictionary.go
@@ -1,0 +1,38 @@
+package main
+
+const (
+	ErrNotFound   = DictionaryErr("could not find the word you were looking for")
+	ErrWordExists = DictionaryErr("cannot add word because it already exists")
+)
+
+type DictionaryErr string
+
+func (e DictionaryErr) Error() string {
+	return string(e)
+}
+
+type Dictionary map[string]string
+
+func (d Dictionary) Search(word string) (string, error) {
+	def, ok := d[word]
+	if !ok {
+		return "", ErrNotFound
+	}
+
+	return def, nil
+}
+
+func (d Dictionary) Add(word, def string) error {
+	_, err := d.Search(word)
+	switch err {
+	case ErrNotFound:
+		d[word] = def
+	case nil:
+		return ErrWordExists
+	default:
+		return err
+
+	}
+
+	return nil
+}
