+++ maps/v7/dictionary_test.go
@@ -1,0 +1,109 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestSearch(t *testing.T) {
+	dictionary := Dictionary{"test": "this is just a test"}
+
+	t.Run("known word", func(t *testing.T) {
+		got, _ := dictionary.Search("test")
+		want := "this is just a test"
+
+		assertStrings(t, got, want)
+	})
+
+	t.Run("unknown word", func(t *testing.T) {
+		_, got := dictionary.Search("unknown")
+
+		assertError(t, got, ErrNotFound)
+	})
+}
+
+func TestAdd(t *testing.T) {
+	t.Run("new word", func(t *testing.T) {
+		dictionary := Dictionary{}
+		word := "test"
+		def := "this is just a test"
+
+		err := dictionary.Add(word, def)
+
+		assertError(t, err, nil)
+		assertDef(t, dictionary, word, def)
+	})
+
+	t.Run("existing word", func(t *testing.T) {
+		word := "test"
+		def := "this is just a test"
+		dictionary := Dictionary{word: def}
+		err := dictionary.Add(word, "new test")
+
+		assertError(t, err, ErrWordExists)
+		assertDef(t, dictionary, word, def)
+	})
+}
+
+func TestUpdate(t *testing.T) {
+	t.Run("existing word", func(t *testing.T) {
+		word := "test"
+		def := "this is just a test"
+		newDef := "new def"
+		dictionary := Dictionary{word: def}
+		err := dictionary.Update(word, newDef)
+
+		assertError(t, err, nil)
+		assertDef(t, dictionary, word, newDef)
+	})
+
+	t.Run("new word", func(t *testing.T) {
+		word := "test"
+		def := "this is just a test"
+		dictionary := Dictionary{}
+
+		err := dictionary.Update(word, def)
+
+		assertError(t, err, ErrWordDoesNotExist)
+	})
+}
+
+func TestDelete(t *testing.T) {
+	word := "test"
+	dictionary := Dictionary{word: "test def"}
+
+	dictionary.Delete(word)
+
+	_, err := dictionary.Search(word)
+	if err != ErrNotFound {
+		t.Errorf("Expected '%s' to be deleted", word)
+	}
+}
+
+func assertStrings(t *testing.T, got, want string) {
+	t.Helper()
+
+	if got != want {
+		t.Errorf("got '%s' want '%s'", got, want)
+	}
+}
+
+func assertError(t *testing.T, got, want error) {
+	t.Helper()
+
+	if got != want {
+		t.Errorf("got error '%s' want '%s'", got, want)
+	}
+}
+
+func assertDef(t *testing.T, dictionary Dictionary, word, def string) {
+	t.Helper()
+
+	got, err := dictionary.Search(word)
+	if err != nil {
+		t.Fatal("should find added word:", err)
+	}
+
+	if def != got {
+		t.Errorf("got '%s' want '%s'", got, def)
+	}
+}
